requirements.txt
matplotlib==3.5.3
numpy==1.21.5
opencv_python==4.5.1.48
Pillow==9.2.0
scikit_image==0.19.3
skimage==0.0
torch==1.11.0
torchvision==0.12.0
